+++ src/core/platform/opengl/opengl_buffer.h
@@ -5,8 +5,8 @@
 struct OpenGLVertexBuffer : mbcore::VertexBuffer
 {
 	OpenGLVertexBuffer(const void* vertices, 
-		const uint32_t count,
-		const uint32_t size,
+		const size_t count,
+		const size_t size,
 		int flags);
 	~OpenGLVertexBuffer();
 
@@ -14,17 +14,17 @@ struct OpenGLVertexBuffer : mbcore::VertexBuffer
 	void* map_buffer_range() const override;
 protected:
 	void initialize(const void* vertices, 
-		const uint32_t count, 
-		const uint32_t size,
+		const size_t count,
+		const size_t size,
 		int flags) override;
 };
 
 struct OpenGLIndexBuffer : mbcore::IndexBuffer
 {
-	OpenGLIndexBuffer(const uint32_t* indices, const uint32_t size);
+	OpenGLIndexBuffer(const uint32_t* indices, const size_t size);
 	~OpenGLIndexBuffer();
 protected:
-	void initialize(const uint32_t* indices, const uint32_t size) override;
+	void initialize(const uint32_t* indices, const size_t size) override;
 };
 
 #endif // !_OPENGL_BUFFER_H
